pyproject.toml
[tool.poetry]
name = "mnist-pytorch"
version = "0.1.0"
description = ""
authors = ["Maxime Boulanger <boumax84@gmail.com>"]
readme = "README.md"
packages = [{include = "mnist_pytorch"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
torchvision = "^0.14.1"
pytorch-lightning = "^1.9.3"
fire = "^0.5.0"
tensorboard = "^2.12.0"
rich = "^13.3.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
pytest = "^7.2.1"
eradicate = "^2.1.0"
pybetter = "^0.4.1"
pylint = "^2.16.2"
jupyter = "^1.0.0"

[tool.poetry.group.api.dependencies]
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
djangorestframework-api-key = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 1
